+++ src/@types/index.d.ts
@@ -44,6 +44,7 @@ export declare interface ParserOptions extends PropertyListOptions {
     resolveUrls?: boolean;
     resolveImport?: boolean;
     cwd?: string;
+    parseColor?: boolean;
     removeDuplicateDeclarations?: boolean;
     computeShorthand?: boolean;
     inlineCssVariables?: boolean;
@@ -122,7 +123,6 @@ export declare interface TransformResult extends ParseResult, RenderResult {
 }
 
 export declare interface ParseTokenOptions extends ParserOptions {
-    parseColor?: boolean;
 }
 
 export declare interface TokenizeResult {