+++ src/@types/token.d.ts
@@ -341,14 +341,16 @@ export declare interface ImportantToken extends BaseToken {
     typ: EnumToken.ImportantTokenType;
 }
 
-export declare type ColorType = 'lit' | 'hex' | 'rgb' | 'rgba' | 'hsl' | 'hsla' | 'hwb' | 'device-cmyk';
+export declare type ColorKind = 'lit' | 'hex' | 'rgb' | 'rgba' | 'hsl' | 'hsla' | 'hwb' | 'device-cmyk';
 
 export declare interface ColorToken extends BaseToken {
 
     typ: EnumToken.ColorTokenType;
     val: string;
-    kin: ColorType;
+    kin: ColorKind;
     chi?: Token[];
+    /* calculated */
+    cal?: 'rel' | 'mix';
 }
 
 export declare interface AttrToken extends BaseToken {