+++ src/lib/parser/parse.ts
@@ -101,6 +101,7 @@ export async function doParse(iterator: string, options: ParserOptions = {}): Pr
             src: '',
             sourcemap: false,
             minify: true,
+            parseColor: true,
             nestingRules: false,
             resolveImport: false,
             resolveUrls: false,
@@ -454,7 +455,7 @@ export async function doParse(iterator: string, options: ParserOptions = {}): Pr
 
                             name = tokens.slice(0, i);
                             value = parseTokens(tokens.slice(i + 1), {
-                                parseColor: true,
+                                parseColor: options.parseColor,
                                 src: options.src,
                                 resolveUrls: options.resolveUrls,
                                 resolve: options.resolve,
@@ -1006,15 +1007,6 @@ export function parseTokens(tokens: Token[], options: ParseTokenOptions = {}) {
                 // @ts-ignore
                 parseTokens(t.chi, t.typ, options);
             }
-            // @ts-ignore
-            // t.chi.forEach(val => {
-            //     if (val.typ == EnumToken.StringTokenType) {
-            //         const slice = val.val.slice(1, -1);
-            //         if ((slice.charAt(0) != '-' || (slice.charAt(0) == '-' && isIdentStart(slice.charCodeAt(1)))) && isIdent(slice)) {
-            //             Object.assign(val, {typ: EnumToken.IdenTokenType, val: slice});
-            //         }
-            //     }
-            // });
 
             let m: number = (<Token[]>t.chi).length;
             let val: Token;
@@ -1197,6 +1189,13 @@ export function parseTokens(tokens: Token[], options: ParseTokenOptions = {}) {
                 t.chi.pop();
             }
 
+            // @ts-ignore
+            if (t.chi.length > 0) {
+
+                // @ts-ignore
+                parseTokens(t.chi, options);
+            }
+
             if (t.typ == EnumToken.FunctionTokenType && (<FunctionToken>t).val == 'calc') {
 
                 for (const {value, parent} of walkValues((<FunctionToken>t).chi)) {
@@ -1240,37 +1239,19 @@ export function parseTokens(tokens: Token[], options: ParseTokenOptions = {}) {
             // @ts-ignore
             if (options.parseColor && t.typ == EnumToken.FunctionTokenType && isColor(t)) {
 
-                // if (isColor) {
                 // @ts-ignore
                 t.typ = EnumToken.ColorTokenType;
                 // @ts-ignore
                 t.kin = t.val;
 
-                t.chi = t.chi.filter((t: Token) => ![EnumToken.WhitespaceTokenType, EnumToken.CommaTokenType, EnumToken.CommentTokenType].includes(t.typ) && !(t.typ == EnumToken.LiteralTokenType && t.val == '/'));
+                if (t.chi[0].typ == EnumToken.IdenTokenType && t.chi[0].val == 'from') {
 
-                // t.chi.length = 0;
-                // @ts-ignore
-                // let m = t.chi.length;
+                    // @ts-ignore
+                    t.cal = 'rel';
+                }
 
-                // while (m-- > 0) {
-                //     // @ts-ignore
-                //     if ([EnumToken.IdenTokenType].concat(trimWhiteSpace).includes(t.chi[m].typ)) {
-                //         // @ts-ignore
-                //         if (t.chi[m + 1]?.typ == EnumToken.WhitespaceTokenType) {
-                //
-                //             // @ts-ignore
-                //             t.chi.splice(m + 1, 1);
-                //         }
-                //         // @ts-ignore
-                //         if (t.chi[m - 1]?.typ == EnumToken.WhitespaceTokenType) {
-                //
-                //             // @ts-ignore
-                //             t.chi.splice(m - 1, 1);
-                //             m--;
-                //         }
-                //     }
-                // }
 
+                t.chi = t.chi.filter((t: Token) => ![EnumToken.WhitespaceTokenType, EnumToken.CommaTokenType, EnumToken.CommentTokenType].includes(t.typ));
                 continue;
             }
 
@@ -1298,8 +1279,6 @@ export function parseTokens(tokens: Token[], options: ParseTokenOptions = {}) {
 
             // @ts-ignore
             if (t.chi.length > 0) {
-                // @ts-ignore
-                parseTokens(t.chi, options);
                 if (t.typ == EnumToken.PseudoClassFuncTokenType && t.val == ':is' && options.minify) {
                     //
                     const count = t.chi.filter(t => t.typ != EnumToken.CommentTokenType).length;
@@ -1321,7 +1300,7 @@ export function parseTokens(tokens: Token[], options: ParseTokenOptions = {}) {
 
             if (t.typ == EnumToken.IdenTokenType) {
                 // named color
-                const value = t.val.toLowerCase();
+                const value: string = t.val.toLowerCase();
 
                 if (value in COLORS_NAMES) {
                     Object.assign(t, {